Extract loader logic to `doTransform`
